+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- ErrorView react component for rendering errors in forms
+- API localization with locale param
+
 ## [0.6.3] - 2019-11-06
 ### Added
 - discord_bot library